% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{contributions}
\alias{contributions}
\title{Documente les contributions à une évolution}
\usage{
contributions(
  df,
  temps = "croissant",
  vart,
  Tglissement = 1,
  seuilpc = 20,
  lang = get_serad_language()
)
}
\arguments{
\item{df}{Le dataframe en entrée}

\item{temps}{"croissant" par défaut si du passé vers le présent. "décroissant" sinon.}

\item{vart}{La variable de temps du dataframe. La première valeur par défaut.}

\item{Tglissement}{Sur combien de périodes faut-il remonter pour calculer l'évolution à 1 par défaut.}

\item{seuilpc}{Un seuil en dessous duquel la contribution (en valeur absolue)
n'est pas prise en compte. 5 \% par défaut.}

\item{lang}{Langue de sortie : "fr" ou "en".}
}
\value{
Un descriptif ordonné des différentes glissements.
}
\description{
Un dataframe en entrée.\cr
Si non indiqué, la date est considérée comme étant la première colonne.
Pas besoin de la formater.\cr
Si non indiqué, on suppose les valeurs les plus anciennes en haut
(temps croissant).\cr
Toutes les autres colonnes sont considérées comme les composantes dont les
contributions doivent être analysées. Elles doivent être numériques.\cr
}
\details{
Fonction écrite sur demande d'une utilisatrice.

Si aucune contribution n'est supérieure à seuilpc en valeur absolue,
contributions() retourne a minima la plus grande contribution de même sens
que la contribution totale.
}
\examples{
col0 = c("Y1T1", "Y1T2", "Y1trim3", "Y1T4","Y2T1","Y2-T2")
col1 = c( 12,      6,      2,         86,     19,     10)
col2 = c(  4,      8,      7,         34,     87,     14)
col3 = c( 10,     20,      3,         66,     90,     54)
col4 = c( 29,     12,      4,         16,     40,     94)
col5 = c( 58,     76,      1,          3,     34,     19)
df1 = data.frame(col0,col1,col2,col3,col4,col5)

contributions(df1)

}
