% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_nom0.R
\name{g_nom0}
\alias{g_nom0}
\title{Evolution décrite de facon nominale, non suivie d'une évolution.}
\usage{
g_nom0(g)
}
\arguments{
\item{g}{L'évolution, en pourcentage.}
}
\value{
L'évolution, par exemple: "une forte hausse".
}
\description{
Evolution décrite de facon nominale, non suivie d'une évolution.
}
\examples{
g_nom0(4)       # une forte hausse
g_nom0(1)       # une hausse
g_nom0(0.4)     # une hausse modérée
g_nom0(0.1)     # une légère hausse
g_nom0(0)       # une stabilité
g_nom0(-0.3)    # une légère baisse
g_nom0(-1)      # une baisse modérée
g_nom0(-4)      # une baisse
g_nom0(-5)      # une forte baisse

}
