% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparaison.R
\name{g_nom_simple}
\alias{g_nom_simple}
\title{Évolution nominale simple}
\usage{
g_nom_simple(x1, x2, lang = get_serad_language())
}
\arguments{
\item{x1}{Le niveau le plus récent.}

\item{x2}{Le niveau le plus ancien.}

\item{lang}{Langue de sortie : "fr" ou "en".}
}
\value{
Une chaîne de caractères.
}
\description{
Décrit l'évolution entre \code{x1} et \code{x2}
sous forme nominale suivie du pourcentage correspondant.
}
