% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_verbe0.R
\name{g_verbe0}
\alias{g_verbe0}
\title{Evoluation verbale ne tenant pas compte de l'acceleration et suivi de "de 'ga'"}
\usage{
g_verbe0(g, sing = 1)
}
\arguments{
\item{g}{L'evolution}

\item{sing}{1 si le sujet du verbe est singulier (défault), 0 sinon}
}
\value{
l'evolution, par exemple: "bondit de 10,0 \%"
}
\description{
Evoluation verbale ne tenant pas compte de l'acceleration et suivi de "de 'ga'"
}
\examples{
g_verbe0(10,1) # bondit de 10,0 \%
g_verbe0(4,1)  # s'accroit de 4,0 \%
g_verbe0(1,0)  # sont en hausse de 1,0 \%
g_verbe0(0.3)  # augmente de 0,3 \%
g_verbe0(0.1)  # s'accroit très légèrement de 0,1 \%
g_verbe0(-0.1) # est stable à 0,1 \%
g_verbe0(-0.3) # diminue légèrement de 0,3 \%
g_verbe0(-1)   # recule légèrement de 1,0 \%
g_verbe0(-4)   # baisse de 4,0 \%
g_verbe0(-20)  # recule de 20,0 \%
g_verbe0(-21)  # chute de 21,0 \%

}
\seealso{
g_verbe
}
