% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_serad_language.R
\name{get_serad_language}
\alias{get_serad_language}
\title{Récupère la langue de serad}
\usage{
get_serad_language()
}
\value{
Une chaîne de caractères indiquant la langue courante ("fr" ou "en").
}
\description{
Retourne la langue actuellement utilisée par \{serad\}.
}
