% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gETa_verbe0.R
\name{gETa_verbe0}
\alias{gETa_verbe0}
\title{Pour une évolution verbale tenant compte de l'acceleration et non suivi de g}
\usage{
gETa_verbe0(g1, g2, sing = 1)
}
\arguments{
\item{g1}{La dernière evolution, exprimé en pourcent}

\item{g2}{L'évolution précédente, exprimée en pourcent}

\item{sing}{1 si le sujet du verbe est singulier (défault), 0 sinon}
}
\value{
Une modalité (un nombre). gETa_verbe0 indiquera concrètement le verbe
}
\description{
Pour une évolution verbale tenant compte de l'acceleration et non suivi de g
}
\examples{
gETa_verbe0(0.049,0.049) #reste stable
gETa_verbe0(0.049,2)     #se stabilise
gETa_verbe0(10,1)        #accélère
gETa_verbe0(4,1,0)       #accélèrent
gETa_verbe0(1,1)         #poursuit sa progression
gETa_verbe0(0.3,1)       #ralentit
gETa_verbe0(0.1,-1)      #repart à la hausse
gETa_verbe0(-0.1,-1)     #poursuit sa baisse
gETa_verbe0(-0.3,-1)     #poursuit sa baisse
gETa_verbe0(-1,-1)       #poursuit sa baisse
gETa_verbe0(-4,-1)       #recule à nouveau
gETa_verbe0(-4,1)        #recule
gETa_verbe0(-20,1)       #se replie fortement
gETa_verbe0(-21,1)       #chute

}
\seealso{
gETa_verbe00
}
