% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gETa_verbe00.R
\name{gETa_verbe00}
\alias{gETa_verbe00}
\title{Pour une évolution verbale tenant compte de l'acceleration et non suivi de g}
\usage{
gETa_verbe00(g1, g2)
}
\arguments{
\item{g1}{La dernière evolution, exprimée en pourcent}

\item{g2}{L'évolution précédente, exprimée en pourcent}
}
\value{
Une modalité (un nombre). gETa_verbe0 indiquera concrètement le verbe
}
\description{
Pour une évolution verbale tenant compte de l'acceleration et non suivi de g
}
\examples{
gETa_verbe00(0.049,0.049) #1
gETa_verbe00(0.049,2)     #2
gETa_verbe00(10,1)        #5
gETa_verbe00(4,1)         #5
gETa_verbe00(1,1)         #7
gETa_verbe00(0.3,1)       #6
gETa_verbe00(0.1,-1)      #3
gETa_verbe00(-0.1,-1)     #11
gETa_verbe00(-0.3,-1)     #11
gETa_verbe00(-1,-1)       #11
gETa_verbe00(-4,-1)       #10
gETa_verbe00(-4,1)        #8
gETa_verbe00(-20,1)       #13
gETa_verbe00(-21,1)       #12

}
\seealso{
gETa_verbe0
}
