% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluriel.R
\name{pluriel}
\alias{pluriel}
\title{Accord au pluriel}
\usage{
pluriel(x, sing = "", plur = "s", lang = get_serad_language())
}
\arguments{
\item{x}{Valeur numérique.}

\item{sing}{Forme utilisée au singulier. Par défaut : chaîne vide.}

\item{plur}{Forme utilisée au pluriel. Par défaut : \code{"s"}.}

\item{lang}{Langue de sortie : \code{"fr"} ou \code{"en"}.}
}
\value{
Une chaîne de caractères correspondant à la forme correctement accordée.
Renvoie \code{NA_character_} lorsque \code{x} vaut \code{NA}.
}
\description{
Retourne une forme singulière ou plurielle selon une valeur numérique
et la langue.
}
\details{
En français, le pluriel s'applique à partir de 2.
En anglais, seul 1 (ou -1) est au singulier.
}
\examples{
pluriel(-7.5)
pluriel(-2)
pluriel(1.4, "chat parle", "chats parlent")
pluriel(-2, "chat parle", "chats parlent")
pluriel(1.97)
pluriel(1.5, "point", "points", lang = "fr")
pluriel(1.5, "point", "points", lang = "en")

}
