% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quelTrim.R
\name{quelTrim}
\alias{quelTrim}
\title{Fournit dynamiquement le trimestre en toutes lettres ou en chiffres}
\usage{
quelTrim(
  trim,
  annee,
  type = c("lettres", "chiffres"),
  majuscule = FALSE,
  exposant = TRUE,
  mois = NULL,
  lang = get_serad_language()
)
}
\arguments{
\item{trim}{Trimestre sous forme numérique : 1, 2, 3 ou 4.}

\item{annee}{Année sur 4 chiffres.}

\item{type}{"lettres" (par défaut) ou "chiffres".}

\item{majuscule}{Indicateur logique : TRUE pour une majuscule initiale
(ex. "Premier"), FALSE sinon.}

\item{exposant}{Indicateur logique : TRUE pour afficher
les exposants Markdown (ex. "1^er^"),
FALSE pour afficher "1er".}

\item{mois}{Si renseigné (entier entre 1 et 12),
remplace \code{trim} en déduisant le trimestre correspondant.}

\item{lang}{Langue de sortie : "fr" ou "en".}
}
\value{
Une chaîne de caractères du type
"troisième trimestre 2023"
ou "3^e^ trimestre 2023".
}
\description{
Retourne une formulation littéraire du trimestre
pour une année donnée.
}
\examples{
quelTrim(3, 2023)
quelTrim(3, 2023, majuscule = TRUE)
quelTrim(3, 2023, type = "chiffres")
quelTrim(1, 2023, type = "chiffres")
quelTrim(999, 2023, type = "chiffres", mois = 8)
quelTrim(3, 2023, type = "chiffres", exposant = FALSE)

}
\seealso{
\code{\link{nextTrim}}, \code{\link{prevTrim}}
}
