% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{s}
\alias{s}
\title{Met au pluriel selon la valeur numérique}
\usage{
s(a, sing = "", plur = "s", seuil = 2)
}
\arguments{
\item{a}{Valeur numérique.}

\item{sing}{Forme au singulier. Par défaut : chaîne vide.}

\item{plur}{Forme au pluriel. Par défaut : "s".}

\item{seuil}{Seuil à partir duquel la forme plurielle
est utilisée (par défaut : 2).}
}
\value{
Une chaîne de caractères correspondant à la forme
correctement accordée. NA si l'entrée est NA.
}
\description{
Retourne la forme singulière ou plurielle
selon la valeur absolue du nombre.
}
\examples{
serad::s(-7.5)
serad::s(-2)
serad::s(1.4, "chat parle", "chats parlent")
serad::s(-2, "chat parle", "chats parlent")
serad::s(1.97)
serad::s(arrondi_tot(1.97))
serad::s(1.97, seuil = 1.95)

}
\seealso{
\url{https://www.dictionnaire-academie.fr/article/QDL057}
}
