% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_serad_language.R
\name{set_serad_language}
\alias{set_serad_language}
\title{Définit la langue de serad}
\usage{
set_serad_language(lang = c("fr", "en"))
}
\arguments{
\item{lang}{Langue de sortie : "fr" ou "en".}
}
\value{
NULL invisiblement.
}
\description{
Modifie la langue utilisée par \{serad\} et réinitialise
les règles de rédaction associées.
}
