% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{a}
\alias{a}
\title{Calcul d'une accélération}
\usage{
a(x1, x2, x3)
}
\arguments{
\item{x1}{Le niveau le plus récent}

\item{x2}{Le niveau le niveau avant x1}

\item{x3}{Le niveau le plus ancien}
}
\value{
L'accélération, exprimée en pourcentage.
}
\description{
Calcul d'une accélération
}
\details{
Pas (encore) de gestion du cas ou g(x2,x3) est nul, mais cela ne devrait pas se rencontrer en pratique
}
\examples{
a(4,2,1) #0
a(6,2,1) #100

}
\seealso{
g format_g
}
