% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{a}
\alias{a}
\title{Calcul d'une accélération}
\usage{
a(x1, x2, x3)
}
\arguments{
\item{x1}{Niveau le plus récent.}

\item{x2}{Niveau précédent.}

\item{x3}{Niveau le plus ancien.}
}
\value{
Un nombre numérique correspondant à l'accélération en pourcentage.
}
\description{
Calcule l'accélération entre trois niveaux successifs en comparant
les taux de variation consécutifs.
}
\details{
L'accélération correspond à la variation du taux entre
\code{(x1, x2)} et \code{(x2, x3)}.
}
\examples{
a(4, 2, 1)  # 0
a(6, 2, 1)  # 100
a(2, 1, 1)  # valeur très élevée si taux précédent proche de zéro

}
\seealso{
\code{\link{g}}, \code{\link{format_g}}
}
