% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparaison.R
\name{alahausse}
\alias{alahausse}
\title{Evolution a la hausse, a la baisse ou inchangee}
\usage{
alahausse(x1, x2, seuil = 0.1)
}
\arguments{
\item{x1}{Niveau le plus recent.}

\item{x2}{Niveau le plus ancien.}

\item{seuil}{Seuil d'egalite en valeur absolue.
Par defaut : 0.1 (soit 0.1 \%).}
}
\value{
Une chaine de caracteres : "a la hausse", "a la baisse"
ou "inchange".
}
\description{
Indique si \code{x1} evolue a la hausse, a la baisse ou reste
inchange par rapport a \code{x2}.
}
\examples{
alahausse(1.004, 1)    # "à la hausse"
alahausse(0.996, 1)    # "à la baisse"
alahausse(1, 1.0004)   # "inchangé"

}
\seealso{
\code{\link{comparaison}}
}
