% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparaison.R
\name{alahausse}
\alias{alahausse}
\title{Évolution à la hausse, à la baisse ou inchangée}
\usage{
alahausse(x1, x2, seuil = 0.1, lang = get_serad_language())
}
\arguments{
\item{x1}{Niveau le plus récent.}

\item{x2}{Niveau le plus ancien.}

\item{seuil}{Seuil d'égalité en valeur absolue.}

\item{lang}{Langue de sortie : "fr" ou "en".}
}
\value{
Une chaîne de caractères.
}
\description{
Indique si \code{x1} évolue à la hausse, à la baisse ou reste
inchangé par rapport à \code{x2}.
}
\seealso{
\code{\link{comparaison}}
}
