% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alahausse.R
\name{alahausse}
\alias{alahausse}
\title{à la hausse //  à la baisse // inchangé}
\usage{
alahausse(x1, x2, seuil = 0.1)
}
\arguments{
\item{x1}{Le niveau le plus récent}

\item{x2}{Le niveau le plus ancien}

\item{seuil}{La limite pour l'égalite (par défaut 0.1, c'est-à-dire 0.1\%)}
}
\value{
à la hausse // inchangé // à la baisse
}
\description{
à la hausse //  à la baisse // inchangé
}
\examples{
alahausse(1.004,1)   # à la hausse
alahausse(0.996,1)   # à la baisse
alahausse(1,1.0004)  # inchangé

}
\seealso{
comparaison
}
