% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrondi_tot.R
\name{arrondi_tot}
\alias{arrondi_tot}
\title{Arrondi arithmetique}
\usage{
arrondi_tot(x, digits = 1)
}
\arguments{
\item{x}{Nombre a arrondir.}

\item{digits}{Entier indiquant le nombre de decimales.
Positif pour les decimales, negatif pour les dizaines, centaines, etc.
Par defaut : 1.}
}
\value{
Un nombre numerique arrondi.
}
\description{
Arrondit un nombre selon la regle arithmetique (0.5 vers le haut),
contrairement a \code{round()} qui utilise l'arrondi bancaire.
}
\examples{
arrondi_tot(1877.85, digits = 0)   # 1878
arrondi_tot(1877.85, digits = 1)   # 1877.9
arrondi_tot(1877.85, digits = -1)  # 1880

}
\seealso{
\code{\link{format_niv}}
}
