% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrondi_tot.R
\name{arrondi_tot}
\alias{arrondi_tot}
\title{Arrondi arithmétique}
\usage{
arrondi_tot(x, digits = 1)
}
\arguments{
\item{x}{Nombre à arrondir.}

\item{digits}{Entier indiquant le nombre de décimales.
Positif pour les décimales, négatif pour les dizaines, centaines, etc.
Par défaut : 1.}
}
\value{
Un nombre numérique arrondi.
}
\description{
Arrondit un nombre selon la règle arithmétique (0.5 vers le haut),
contrairement à \code{round()} qui utilise l'arrondi bancaire.
}
\examples{
arrondi_tot(1877.85, digits = 0)   # 1878
arrondi_tot(1877.85, digits = 1)   # 1877.9
arrondi_tot(1877.85, digits = -1)  # 1880

}
\seealso{
\code{\link{format_niv}}
}
