% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrondi_tot.R
\name{arrondi_tot}
\alias{arrondi_tot}
\title{Arrondit un nombre.}
\usage{
arrondi_tot(x, digit = 1)
}
\arguments{
\item{x}{Le nombre à arrondir.}

\item{digit}{Un entier positif (pour arrondir après la virgule) ou négatif (pour arrondir à la centaine ou au millier par exemple). Par défaut digit = 1.}
}
\value{
Le nombre arrondi.
}
\description{
Arrondit un nombre.
}
\details{
Notoirement utile car round(0.35) donne...0.3.
Merci a Chloe Pariset et Gaetan Guillermin pour cette fonction.
}
\examples{
arrondi_tot(1877.85,digit=0)  #1878
arrondi_tot(1877.85,digit=1)  #1877.9
arrondi_tot(1877.85,digit=2)  #1877.85
arrondi_tot(1877.85,digit=-1) #1880
arrondi_tot(1877.85,digit=-2) #1900

}
\seealso{
format_niv
}
