% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audessus.R
\name{audessus}
\alias{audessus}
\title{au-dessus ou en dessous}
\usage{
audessus(x1, x2)
}
\arguments{
\item{x1}{Le niveau le plus récent}

\item{x2}{Le niveau le plus ancien}
}
\value{
au-dessus // en dessous
}
\description{
au-dessus ou en dessous
}
\examples{
audessus(1.04,1)    # au-dessus
audessus(1.04,1.04) # au-dessus
audessus(0.96,1)    # en dessous

}
\seealso{
comparaison
}
