% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparaison.R
\name{comparaison}
\alias{comparaison}
\title{comparaison : fonction générique basée sur x1 et x2}
\usage{
comparaison(
  x1,
  x2,
  hausse0,
  egalite0,
  baisse0,
  seuil = 0.1,
  param = 0,
  hausse1 = hausse0,
  egalite1 = egalite0,
  baisse1 = baisse0
)
}
\arguments{
\item{x1}{Le niveau le plus récent}

\item{x2}{le niveau le plus ancien}

\item{hausse0}{Mot si hausse}

\item{egalite0}{Mot si egalite}

\item{baisse0}{Mot si baisse}

\item{seuil}{La limite pour l'egalite (par defaut 0.1\%)}

\item{param}{Un parametre supplementaire égal à 0 ou 1 (par exemple pour singulier)}

\item{hausse1}{Formulation différente si hausse, dépendant de param}

\item{egalite1}{Formulation différente si egalite, dépendant de param}

\item{baisse1}{Formulation différente si baisse, dépendant de param}
}
\value{
Un mot.
}
\description{
comparaison : fonction générique basée sur x1 et x2
}
\details{
Utilise comparaison_taux.R
}
\examples{
comparaison(1.04,1,"augmente","reste stable","diminue")             #augmente
comparaison(0.9991,1,"augmente","reste stable","diminue")           #reste stable
comparaison(0.999,1,"augmente","reste stable","diminue")            #diminue
comparaison(0.9991,1,"augmente","reste stable","diminue",seuil = 0) #diminue
comparaison(1,1,"augmente","reste égal","diminue",seuil = 0)        #augmente

}
\seealso{
comparaison_taux alahausse audessus davantage depasse g_nom_simple
}
