% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparaison_taux.R
\name{comparaison_taux}
\alias{comparaison_taux}
\title{Comparaison d'une variation à un seuil}
\usage{
comparaison_taux(
  g,
  hausse_defaut,
  egalite_defaut,
  baisse_defaut,
  seuil = 0.1,
  alt = 0,
  hausse_alt = hausse_defaut,
  egalite_alt = egalite_defaut,
  baisse_alt = baisse_defaut
)
}
\arguments{
\item{g}{Variation exprimée en pourcentage
(5 signifie 5 \%, 0.1 signifie 0.1 \%).}

\item{hausse_defaut}{Mot si hausse (forme par défaut).}

\item{egalite_defaut}{Mot si égalité (forme par défaut).}

\item{baisse_defaut}{Mot si baisse (forme par défaut).}

\item{seuil}{Limite pour l'égalité (0.1 par défaut).}

\item{alt}{Paramètre supplémentaire égal à 0 ou 1
(par exemple pour distinguer singulier/pluriel).}

\item{hausse_alt}{Formulation alternative si alt = 1.}

\item{egalite_alt}{Formulation alternative si alt = 1.}

\item{baisse_alt}{Formulation alternative si alt = 1.}
}
\value{
Une chaîne de caractères correspondant à la modalité choisie.
}
\description{
Comparaison d'une variation à un seuil
}
\details{
Fonction interne utilisée par \code{\link{comparaison}}.
}
\examples{
comparaison_taux(5, "augmente", "reste stable", "diminue")
comparaison_taux(0.05, "augmente", "reste stable", "diminue")
comparaison_taux(0, "as", "bs", "cs", seuil = 0, alt = 1,
                 hausse_alt = "a", egalite_alt = "b", baisse_alt = "c")

}
\seealso{
\code{\link{comparaison}}
}
