% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparaison.R
\name{davantage}
\alias{davantage}
\title{Davantage ou moins}
\usage{
davantage(x1, x2)
}
\arguments{
\item{x1}{Niveau le plus recent.}

\item{x2}{Niveau le plus ancien.}
}
\value{
Une chaine de caracteres : "davantage" ou "moins".
}
\description{
Indique si \code{x1} est davantage ou moins eleve que \code{x2}.
}
\examples{
davantage(1.04, 1)     # "davantage"
davantage(1.04, 1.04)  # "davantage"
davantage(0.96, 1)     # "moins"

}
\seealso{
\code{\link{comparaison}}
}
