% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparaison.R
\name{davantage}
\alias{davantage}
\title{Davantage ou moins}
\usage{
davantage(x1, x2, lang = get_serad_language())
}
\arguments{
\item{x1}{Niveau le plus récent.}

\item{x2}{Niveau le plus ancien.}

\item{lang}{Langue de sortie : "fr" ou "en".}
}
\value{
Une chaîne de caractères.
}
\description{
Indique si \code{x1} est davantage ou moins élevé que \code{x2}.
}
\seealso{
\code{\link{comparaison}}
}
