% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/davantage.R
\name{davantage}
\alias{davantage}
\title{davantage ou moins}
\usage{
davantage(x1, x2)
}
\arguments{
\item{x1}{Le niveau le plus récent}

\item{x2}{Le niveau le plus ancien}
}
\value{
davantage ou moins
}
\description{
davantage ou moins
}
\examples{
davantage(1.04,1)    # davantage
davantage(1.04,1.04) # davantage
davantage(0.96,1)    # moins

}
\seealso{
comparaison
}
