% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparaison.R
\name{depasse}
\alias{depasse}
\title{Verbe pour exprimer le sens de l'evolution}
\usage{
depasse(x1, x2, sing = FALSE)
}
\arguments{
\item{x1}{Le niveau le plus recent.}

\item{x2}{Le niveau le plus ancien.}

\item{sing}{Indicateur logique : FALSE si le sujet est pluriel,
TRUE s'il est singulier (par defaut : pluriel).}
}
\value{
Une chaine de caracteres :
"excedent"/"excede",
"sont au niveau de"/"est au niveau de",
ou "sont en dessous de"/"est en dessous de".
}
\description{
Indique si \code{x1} excede, est au niveau de ou est en dessous de
\code{x2}, en tenant compte du nombre grammatical.
}
\examples{
depasse(1.04, 1)               # "excèdent"
depasse(1.04, 1, sing = TRUE)  # "excède"
depasse(0.96, 1)               # "sont en dessous de"
depasse(0.9991, 1)             # "sont au niveau de"

}
\seealso{
\code{\link{comparaison}}
}
