% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparaison.R
\name{depasse}
\alias{depasse}
\title{Verbe pour exprimer le sens de l'évolution}
\usage{
depasse(x1, x2, sing = FALSE, lang = get_serad_language())
}
\arguments{
\item{x1}{Le niveau le plus récent.}

\item{x2}{Le niveau le plus ancien.}

\item{sing}{Indicateur logique : FALSE si le sujet est pluriel,
TRUE s'il est singulier (par défaut : pluriel).}

\item{lang}{Langue de sortie : "fr" ou "en".}
}
\value{
Une chaîne de caractères.
}
\description{
Indique si \code{x1} excède, est au niveau de ou est en dessous de
\code{x2}, en tenant compte du nombre grammatical.
}
\seealso{
\code{\link{comparaison}}
}
