% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparaison.R
\name{depasse}
\alias{depasse}
\title{verbe pour exprimer le sens de l'évolution, pouvant gérer si le nom est singulier ou pluriel}
\usage{
depasse(x1, x2, sing = 0)
}
\arguments{
\item{x1}{Le niveau le plus récent}

\item{x2}{Le niveau le plus ancien}

\item{sing}{0 si le nom est un pluriel, 1 si un singulier. Par défaut un pluriel.}
}
\value{
'excéde//-nt' ou 'est//sont au niveau de' ou 'est//sont en dessous de'
}
\description{
verbe pour exprimer le sens de l'évolution, pouvant gérer si le nom est singulier ou pluriel
}
\examples{
depasse(1.04,1)          # excédent
depasse(1.04,1,sing=1)   # excéde
depasse(0.96,1)          # sont en dessous de
depasse(0.9991,1)        # sont au niveau de

}
\seealso{
comparaison
}
