% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_delta.R
\name{format_delta}
\alias{format_delta}
\title{Formatage des variations en niveau}
\usage{
format_delta(y, signe = TRUE, detail = getOption("serad")$arrondi_niv)
}
\arguments{
\item{y}{La difference a formater.}

\item{signe}{Indicateur logique : TRUE pour afficher le signe,
FALSE sinon (par defaut : TRUE).}

\item{detail}{Precision d'arrondi. Par defaut, on utilise
getOption("serad")$arrondi_niv.}
}
\value{
Une chaine de caracteres correspondant a la variation formatee.
}
\description{
Formate une difference numerique selon les regles d'arrondi
et d'affichage utilisees dans le package.
}
\examples{
format_delta(365484)           # "+365 500"
format_delta(365484, FALSE)    # "365 500"
format_delta(-365484)          # "-365 500"
format_delta(-365484, FALSE)   # "365 500"

}
\seealso{
\code{\link{arrondi_tot}}, \code{\link{format_niv}}
}
