% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_delta.R
\name{format_delta}
\alias{format_delta}
\title{Formatage des variations en niveau}
\usage{
format_delta(
  y,
  signe = TRUE,
  detail = getOption("serad")$arrondi_niv,
  lang = get_serad_language()
)
}
\arguments{
\item{y}{La différence à formater.}

\item{signe}{Indicateur logique : TRUE pour afficher le signe,
FALSE sinon (par défaut : TRUE).}

\item{detail}{Précision d'arrondi. Par défaut, on utilise
getOption("serad")$arrondi_niv.}

\item{lang}{Langue de sortie : "fr" ou "en".}
}
\value{
Une chaîne de caractères correspondant à la variation formatée.
}
\description{
Formate une différence numérique selon les règles d'arrondi
et d'affichage utilisées dans le package.
}
\examples{
format_delta(365484)              # "+365 500"
format_delta(365484, FALSE)       # "365 500"
format_delta(-365484)             # "-365 500"
format_delta(-365484, FALSE)      # "365 500"
format_delta(365484, lang = "en") # "+365,500"

}
\seealso{
\code{\link{arrondi_tot}}, \code{\link{format_niv}}
}
