% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_delta.R
\name{format_delta}
\alias{format_delta}
\title{Formatage Dares des variations (exprimées en nombre, avec signe)}
\usage{
format_delta(y, signe = 1, detail = -2)
}
\arguments{
\item{y}{La différence à formater.}

\item{signe}{1 si on veut le signe, 0 sinon. Par défaut signe = 1.}

\item{detail}{Par défaut -2 pour arrondir à la centaine.}
}
\value{
Le niveau formaté.
}
\description{
Formatage Dares des variations (exprimées en nombre, avec signe)
}
\examples{
format_delta(365484)       #+365 500
format_delta(365484,0)     #365 500
format_delta(-365484)      #−365 500
format_delta(-365484,0,-1) #365 480

}
\seealso{
arrondi_tot
}
