% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_g.R
\name{format_g}
\alias{format_g}
\title{Formatage des variations en pourcentage}
\usage{
format_g(y, signe = TRUE, detail = getOption("serad")$arrondi_pourcent)
}
\arguments{
\item{y}{La variation a formater.}

\item{signe}{Indicateur logique : TRUE pour afficher le signe,
FALSE pour le retirer (par defaut : TRUE).}

\item{detail}{Nombre de chiffres apres la virgule.
Par defaut, on utilise getOption("serad")$arrondi_pourcent.}
}
\value{
Une chaine de caracteres correspondant a la variation formatee
(ex. "+5,4 \%", "-5,4 \%").
}
\description{
Formate une variation exprimee en pourcentage selon les regles
d'arrondi et d'affichage du package.
}
\details{
Le symbole "moins" peut etre personnalise via
getOption("serad")$moins.
}
\examples{
format_g(5.3654, signe = FALSE)  # "5,4 \%"
format_g(5.3654)                 # "+5,4 \%"
format_g(-5.3654, FALSE)         # "5,4 \%"
format_g(-5.3654)                # "-5,4 \%"
format_g(-5.3654, detail = 2)    # "-5,37 \%"
format_g(0.35)                   # "+0,4 \%"

}
\seealso{
\code{\link{g}}, \code{\link{arrondi_tot}}
}
