% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_g.R
\name{format_g}
\alias{format_g}
\title{Formatage Dares des variations, exprimées en pourcentage ('g')}
\usage{
format_g(y, signe = 1)
}
\arguments{
\item{y}{La variation à formater.}

\item{signe}{1 si on veut le signe, 0 sinon. Par défaut signe = 1.}
}
\value{
La variation mise en forme.
}
\description{
signe = 0 permet de retirer le signe, que ce soit un + ou un - (par exemple pour dire: diminue de 5 \\%).
}
\details{
Le symbole 'moins' n'est pas correctement rendu en Word lorsque utilisé avec cat. (on obtient le tiret du milieu)
}
\examples{
format_g(5.3654,0)  # "5,4 \%"
format_g(5.3654,1)  # "+5,4 \%"
format_g(-5.3654,0) # "5,4 \%"
format_g(-5.3654,1) # "−5,4 \%"
format_g(-5.3654)   # "−5,4 \%"
}
\seealso{
g
}
