% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_niv.R
\name{format_niv}
\alias{format_niv}
\title{Formatage des niveaux en nombres}
\usage{
format_niv(y, detail = getOption("serad")$arrondi_niv)
}
\arguments{
\item{y}{Le niveau a formater.}

\item{detail}{Precision d'arrondi. Par defaut, on utilise
getOption("serad")$arrondi_niv.}
}
\value{
Une chaine de caracteres correspondant au niveau formate.
}
\description{
Formate un niveau numerique selon les regles d'arrondi
definies dans le package.
}
\examples{
format_niv(365484)  # "365 500"

}
\seealso{
\code{\link{arrondi_tot}}
}
