% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_niv.R
\name{format_niv}
\alias{format_niv}
\title{Formatage Dares des niveaux, exprimés en nombres}
\usage{
format_niv(y, detail)
}
\arguments{
\item{y}{Le niveau à formater.}

\item{detail}{Par défaut -2 pour arrondir à la centaine.}
}
\value{
Le niveau formaté
}
\description{
Formatage Dares des niveaux, exprimés en nombres
}
\examples{
format_niv(365484)  # 365 500

#' @details Pour changer le comportement par défaut d'arrondi, modifier getOption("serad")$arrondi_niv

}
\seealso{
arrondi_tot
}
