% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_niv.R
\name{format_niv}
\alias{format_niv}
\title{Formatage des niveaux en nombres}
\usage{
format_niv(
  y,
  detail = getOption("serad")$arrondi_niv,
  lang = get_serad_language()
)
}
\arguments{
\item{y}{Le niveau à formater.}

\item{detail}{Précision d'arrondi. Par défaut, on utilise
getOption("serad")$arrondi_niv.}

\item{lang}{Langue de sortie : "fr" ou "en".}
}
\value{
Une chaîne de caractères correspondant au niveau formaté.
}
\description{
Formate un niveau numérique selon les règles d'arrondi
définies dans le package.
}
\examples{
format_niv(365484)                # "365 500"
format_niv(365484, lang = "en")   # "365,500"

}
\seealso{
\code{\link{arrondi_tot}}
}
