% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pts.R
\name{format_pts}
\alias{format_pts}
\title{Formatage des variations en points}
\usage{
format_pts(
  y,
  signe = TRUE,
  detail = getOption("serad")$arrondi_pourcent,
  abrev = FALSE
)
}
\arguments{
\item{y}{La variation a formater.}

\item{signe}{Indicateur logique : TRUE pour afficher le signe,
FALSE pour le retirer (par defaut : TRUE).}

\item{detail}{Nombre de chiffres apres la virgule.
Par defaut, on utilise getOption("serad")$arrondi_pourcent.}

\item{abrev}{Indicateur logique : TRUE pour utiliser l'abreviation
("pt"/"pts"), FALSE pour afficher "point(s)" (par defaut : FALSE).}
}
\value{
Une chaine de caracteres correspondant a la variation formatee
(ex. "+5,4 points", "+5,4 pts").
}
\description{
Formate une variation exprimee en points selon les regles
d'arrondi et d'affichage du package.
}
\details{
Le symbole "moins" peut etre personnalise via
getOption("serad")$moins.
}
\examples{
format_pts(5.3654, signe = FALSE)   # "5,4 points"
format_pts(1.3654, signe = FALSE)   # "1,4 point"
format_pts(5.3654, abrev = TRUE)    # "+5,4 pts"
format_pts(-5.3654, FALSE)          # "5,4 points"
format_pts(-5.3654)                 # "-5,4 points"
format_pts(-5.3654, detail = 2)     # "-5,37 points"
format_pts(0.35)                    # "+0,4 point"

}
\seealso{
\code{\link{format_g}}
}
