% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pts.R
\name{format_pts}
\alias{format_pts}
\title{Formatage des variations, exprimées en points}
\usage{
format_pts(y, signe = 1, detail, abrev = 0)
}
\arguments{
\item{y}{La variation à formater.}

\item{signe}{1 si on veut le signe, 0 sinon. Par défaut signe = 1.}

\item{detail}{Combien de chiffre après la virgule. Par défaut 1.}

\item{abrev}{1 si on préfère l'abréviation. Par défaut abrev = 0.}
}
\value{
La variation mise en forme.
}
\description{
signe = 0 permet de retirer le signe, que ce soit un + ou un - (par exemple pour dire: diminue de 5 points).
}
\details{
Le symbole 'moins' n'est pas correctement rendu en Word lorsque utilisé avec cat. (on obtient le tiret du milieu)

Pour changer le comportement par défaut d'arrondi, modifier getOption("serad")$arrondi_pourcent
}
\examples{
format_pts(5.3654,signe =0)  # "5,4 points"
format_pts(1.3654,signe =0)  # "1,4 point"
format_pts(5.3654,abrev = 1) # "+5,4 pts"
format_pts(5.3654,1)  # "+5,4 points"
format_pts(-5.3654,0) # "5,4 points"
format_pts(-5.3654,1) # "−5,4 points"
format_pts(-5.3654)   # "−5,4 points"
format_pts(-5.3654,detail = 2)   # "−5,37 points"
format_pts(0.35)      # "+0.4 point"
}
\seealso{
format_g
}
