% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{g}
\alias{g}
\title{Calcul d'une variation}
\usage{
g(x1, x2)
}
\arguments{
\item{x1}{Le niveau le plus récent}

\item{x2}{Le niveau le plus ancien}
}
\value{
La variation, exprimée en pourcentage (si x1=2.x1, cela retourne 100).
}
\description{
Calcul d'une variation
}
\details{
Pas (encore) de gestion du cas où x2 est nul.
}
\examples{
g(2,1)  #100

}
\seealso{
format_g
}
