% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{g}
\alias{g}
\title{Calcul d'une variation}
\usage{
g(x1, x2, eps = 1e-08)
}
\arguments{
\item{x1}{Le niveau le plus récent}

\item{x2}{Le niveau le plus ancien}

\item{eps}{Gestion du cas du denominateur nul}
}
\value{
La variation, exprimée en pourcentage (si x1=2.x1, cela retourne 100).
}
\description{
Calcul d'une variation
}
\details{
Dans le cas improbable où x2 = 0, on considère qu'il vaut 0.00000001.
C'est une valeur paramétrable avec eps. L'idée est de toujours sortir quelque chose.
Un message d'avis apparait dans la console.
}
\examples{
g(2,1)  #100
g(2,0)  #2e+10 et message d'avis

}
\seealso{
format_g
}
