% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{g}
\alias{g}
\title{Calcul d'une variation relative}
\usage{
g(x1, x2, eps = 1e-08)
}
\arguments{
\item{x1}{Niveau le plus récent.}

\item{x2}{Niveau le plus ancien.}

\item{eps}{Valeur utilisée à la place de \code{x2} lorsque
\code{x2 = 0}, afin d'éviter une division par zéro.
Par défaut : \code{1e-8}.}
}
\value{
Une valeur numérique correspondant à la variation en pourcentage
(par exemple, si x1 = 2 * x2, la fonction retourne 100).
}
\description{
Calcule la variation relative entre \code{x1} et \code{x2},
exprimée en pourcentage.
}
\details{
Si \code{x2 = 0}, la valeur epsilon définie par
\code{getOption("serad")$eps} est utilisée afin d'éviter une division
par zéro. Un message d'avertissement est émis dans ce cas.

Il est possible de modifier cette valeur :

\code{
serad0 <- getOption("serad")
serad0$eps <- 0
options(serad = serad0)
}
}
\examples{
g(2, 1)  # 100
g(2, 0)  # valeur très élevée et avertissement

}
\seealso{
\code{\link{format_g}}
}
