% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_nom.R
\name{g_nom}
\alias{g_nom}
\title{Evolution nominale non suivie d'une valeur}
\usage{
g_nom(x1, x2, evolution = c("pourcents", "points"), titre = FALSE)
}
\arguments{
\item{x1}{Le niveau le plus recent.}

\item{x2}{Le niveau le plus ancien.}

\item{evolution}{Type d'evolution :
"pourcents" (variation relative, par defaut) ou "points".}

\item{titre}{Indicateur logique : TRUE pour supprimer l'article
initial et mettre une majuscule, notamment en debut de titre.}
}
\value{
Une chaine de caracteres correspondant a la formulation nominale
retenue (par exemple : "une forte hausse").
}
\description{
Decrit une evolution sous forme nominale a partir de deux niveaux,
sans ajouter la valeur de variation.
}
\details{
La fonction calcule d'abord une evolution a partir de \code{x1}
et \code{x2} :
\itemize{
\item si \code{evolution = "pourcents"}, elle utilise
\code{\link{g}} ;
\item si \code{evolution = "points"}, elle calcule \code{x1 - x2}.
}

La valeur obtenue est ensuite transmise a \code{\link{g_nom_taux}},
qui determine la formulation a partir de la table
\code{getOption("serad")$evo_simple}.
}
\examples{
g_nom(1.04, 1)
g_nom(1.01, 1)
g_nom(1.004, 1)
g_nom(1.001, 1)
g_nom(1, 1)
g_nom(0.997, 1)
g_nom(0.95, 1)
g_nom(0.95, 1, evolution = "points")

}
\seealso{
\code{\link{g_nom_taux}}, \code{\link{g}}
}
