% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_nom.R
\name{g_nom}
\alias{g_nom}
\title{Evolution décrite de facon nominale, non suivie d'une évolution.}
\usage{
g_nom(x1, x2)
}
\arguments{
\item{x1}{Le niveau le plus récent}

\item{x2}{le niveau le plus ancien}
}
\value{
L'évolution, par exemple: "une forte hausse".
}
\description{
Evolution décrite de facon nominale, non suivie d'une évolution.
}
\examples{
g_nom(1.04,1)  # une forte hausse
g_nom(1.01,1)  # une hausse
g_nom(1.004,1) # une hausse modérée
g_nom(1.001,1) # une légère hausse
g_nom(1,1)     # une stabilité
g_nom(0.997,1) # une légère baisse
g_nom(0.99,1)  # une baisse modérée
g_nom(0.96,1)  # une baisse
g_nom(0.95,1)  # une forte baisse

}
