% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparaison.R
\name{g_nom_simple}
\alias{g_nom_simple}
\title{Evolution décrite (simplement) de facon nominale, suivie d'une évolution}
\usage{
g_nom_simple(x1, x2)
}
\arguments{
\item{x1}{Le niveau le plus récent}

\item{x2}{Le niveau le plus ancien}
}
\value{
en hausse de//en baisse de + pourcentage
}
\description{
Evolution décrite (simplement) de facon nominale, suivie d'une évolution
}
\examples{
g_nom_simple(3,1) # en hausse de 200,0 \%
g_nom_simple(3,5) # en baisse de 40,0 \%

}
