% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_nom_taux.R
\name{g_nom_taux}
\alias{g_nom_taux}
\title{Évolution nominale d'un taux}
\usage{
g_nom_taux(g, titre = FALSE, lang = get_serad_language())
}
\arguments{
\item{g}{L'évolution en pourcentage.}

\item{titre}{Indicateur logique : TRUE pour supprimer l'article
initial et mettre une majuscule, notamment en début de titre.}

\item{lang}{Langue de sortie : "fr" ou "en".}
}
\value{
Une chaîne de caractères décrivant l'évolution
(par exemple : "une forte hausse", "une stabilité").
}
\description{
Décrit une évolution exprimée en pourcentage sous forme nominale
(par exemple : "une forte hausse").
}
\details{
La fonction sélectionne, dans la table
\code{getOption("serad")$evo_simple}, la ligne dont la condition est vérifiée
par la valeur de \code{g}.

La table \code{evo_simple} doit contenir une colonne \code{condition}, composée de
chaînes de caractères évaluables par R, par exemple :
\code{"g >= -0.10 & g <= 0.10"}.

Les conditions doivent être disjointes : pour une valeur donnée de \code{g},
une seule condition doit être vraie.

Une catégorie est considérée comme une stabilité si sa condition est
également vérifiée pour \code{g = 0}.

La fonction renvoie ensuite la colonne \code{nom} correspondante.
Si \code{titre = TRUE}, l'article initial est supprimé et la première lettre
restante est mise en majuscule.
}
\section{Personnalisation}{

Les formulations utilisées par cette fonction proviennent de la table
\code{getOption("serad")$evo_simple}.

Pour modifier les conditions ou les libellés, voir
\code{\link{init_serad}}.
}

\examples{
g_nom_taux(4)
g_nom_taux(1)
g_nom_taux(0.4)
g_nom_taux(0.1)
g_nom_taux(0)
g_nom_taux(-0.1)
g_nom_taux(-0.3)
g_nom_taux(-1)
g_nom_taux(-4)
g_nom_taux(-5)

}
\seealso{
\code{\link{g_verbe_taux}},
\code{\link{init_serad}}
}
