% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_nom_taux.R
\name{g_nom_taux}
\alias{g_nom_taux}
\title{Evolution nominale d'un taux}
\usage{
g_nom_taux(g, titre = FALSE)
}
\arguments{
\item{g}{L'evolution en pourcentage.}

\item{titre}{Indicateur logique : TRUE pour supprimer l'article
initial et mettre une majuscule, notamment en debut de titre.}
}
\value{
Une chaine de caracteres decrivant l'evolution
(par exemple : "une forte hausse", "une stabilite").
}
\description{
Decrit une evolution exprimee en pourcentage sous forme nominale
(par exemple : "une forte hausse").
}
\details{
La fonction selectionne d'abord, dans la table
\code{getOption("serad")$evo_simple}, la premiere ligne dont
le seuil est inferieur strictement a \code{g}.

La fonction renvoie alors la colonne \code{nom} correspondante.
Si \code{titre = TRUE}, l'article initial est supprime et la
premiere lettre restante est mise en majuscule.
}
\examples{
g_nom_taux(4)
g_nom_taux(1)
g_nom_taux(0.4)
g_nom_taux(0.1)
g_nom_taux(0)
g_nom_taux(-0.3)
g_nom_taux(-1)
g_nom_taux(-4)
g_nom_taux(-5)

}
\seealso{
\code{\link{g_verbe_taux}}
}
