% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_nom_taux.R
\name{g_nom_taux}
\alias{g_nom_taux}
\title{Evolution décrite de facon nominale, non suivie d'une évolution.}
\usage{
g_nom_taux(g)
}
\arguments{
\item{g}{L'évolution, en pourcentage.}
}
\value{
L'évolution, par exemple: "une forte hausse".
}
\description{
Evolution décrite de facon nominale, non suivie d'une évolution.
}
\examples{
g_nom_taux(4)       # une forte hausse
g_nom_taux(1)       # une hausse
g_nom_taux(0.4)     # une hausse modérée
g_nom_taux(0.1)     # une légère hausse
g_nom_taux(0)       # une stabilité
g_nom_taux(-0.3)    # une légère baisse
g_nom_taux(-1)      # une baisse modérée
g_nom_taux(-4)      # une baisse
g_nom_taux(-5)      # une forte baisse


}
