% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gETa_verbe.R
\name{gETa_verbe}
\alias{gETa_verbe}
\title{Pour une évolution verbale tenant compte de l'acceleration et non suivi de g}
\usage{
gETa_verbe(x1, x2, x3, sing = 1)
}
\arguments{
\item{x1}{Le niveau le plus récent}

\item{x2}{Le niveau précédent}

\item{x3}{Le niveau le plus ancien}

\item{sing}{1 si le sujet du verbe est singulier (défault), 0 sinon}
}
\value{
Le verbe.
}
\description{
Pour une évolution verbale tenant compte de l'acceleration et non suivi de g
}
\examples{
gETa_verbe(1.00049,1,0.9996) #reste stable
gETa_verbe(1.00049,1,0.981)  #se stabilise
gETa_verbe(1.1,1,0.99)       #accélère
gETa_verbe(1.04,1,0.99)      #accélère
gETa_verbe(1.01,1,0.99)      #poursuit sa progression
gETa_verbe(1.003,1,0.99,0)   #ralentissent
gETa_verbe(1.001,1,1.01)     #repart à la hausse
gETa_verbe(0.999,1,1.01)     #poursuit sa baisse
gETa_verbe(0.99,1,1.01)      #poursuit sa baisse // recule à nouveau
gETa_verbe(0.96,1,1.01)      #poursuit sa baisse // recule à nouveau
gETa_verbe(0.96,1,0.99)      #recule
gETa_verbe(0.8,1,0.99)       #se replie fortement
gETa_verbe(0.79,1,0.99)      #chute

}
\seealso{
gETa_verbe0
}
