% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gETa_verbe00.R
\name{gETa_verbe00}
\alias{gETa_verbe00}
\title{Pour une évolution verbale tenant compte de l'acceleration et non suivi de g}
\usage{
gETa_verbe00(g1, g2)
}
\arguments{
\item{g1}{La dernière evolution, exprimée en pourcent}

\item{g2}{L'évolution précédente, exprimée en pourcent}
}
\value{
Une modalité (un nombre). gETa_verbe0 indiquera concrètement le verbe
}
\description{
Pour une évolution verbale tenant compte de l'acceleration et non suivi de g
}
\seealso{
gETa_verbe_taux
}
