% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gETa_verbe_taux.R
\name{gETa_verbe_taux}
\alias{gETa_verbe_taux}
\title{Pour une évolution verbale tenant compte de l'acceleration et non suivi de g}
\usage{
gETa_verbe_taux(g1, g2, sing = 1)
}
\arguments{
\item{g1}{La dernière evolution, exprimé en pourcent}

\item{g2}{L'évolution précédente, exprimée en pourcent}

\item{sing}{1 si le sujet du verbe est singulier (défault), 0 sinon}
}
\value{
Une modalité (un nombre). gETa_verbe0 indiquera concrètement le verbe
}
\description{
Pour une évolution verbale tenant compte de l'acceleration et non suivi de g
}
\examples{
gETa_verbe_taux(0.049,0.049) #reste stable
gETa_verbe_taux(0.049,2)     #se stabilise
gETa_verbe_taux(10,1)        #accélère
gETa_verbe_taux(4,1,0)       #accélèrent
gETa_verbe_taux(1,1)         #poursuit sa progression
gETa_verbe_taux(0.3,1)       #ralentit
gETa_verbe_taux(0.1,-1)      #repart à la hausse // se redresse
gETa_verbe_taux(-0.1,-1)     #poursuit sa baisse // continue à baisser
gETa_verbe_taux(-0.3,-1)     #poursuit sa baisse // continue à baisser
gETa_verbe_taux(-1,-1)       #poursuit sa baisse // continue à baisser
gETa_verbe_taux(-4,-1)       #recule à nouveau // poursuit son recul
gETa_verbe_taux(-4,1)        #recule
gETa_verbe_taux(-20,1)       #se replie fortement
gETa_verbe_taux(-21,1)       #chute


}
\seealso{
gETa_verbe00
}
