% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quelTrim.R
\name{nextMois}
\alias{nextMois}
\title{Fournit dynamiquement le mois suivant}
\usage{
nextMois(mois, annee, type = c("Annee", "autres"), majuscule = FALSE, k = 1)
}
\arguments{
\item{mois}{Mois sous forme numerique : 1 a 12.}

\item{annee}{Annee sur 4 chiffres.}

\item{type}{"Annee" (par defaut, ex. "fevrier 2023")
ou "autres" (ex. "fevrier").}

\item{majuscule}{Indicateur logique : TRUE pour une majuscule
initiale (ex. "Fevrier"), FALSE sinon.}

\item{k}{Decalage en nombre de mois.
Par defaut : 1 (mois suivant).}
}
\value{
Une chaine de caracteres correspondant au mois
suivant (par exemple : "avril 2023").
}
\description{
Retourne la formulation du mois situe k periodes
apres le mois indique.
}
\examples{
nextMois(3, 2023)
nextMois(12, 2023)
nextMois(12, 2023, k = 2)

}
\seealso{
\code{\link{quelMois}}, \code{\link{prevMois}}, \code{\link{nextTrim}}
}
