% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quelTrim.R
\name{nextMois}
\alias{nextMois}
\title{Fournit dynamiquement le trimestre suivant sous forme littéraire}
\usage{
nextMois(mois, annee, type = "Annee", majuscule = 0, k = 1)
}
\arguments{
\item{mois}{Le mois sous la forme d'un chiffre : 1..12}

\item{annee}{Un nombre positif en 4 chiffres}

\item{type}{Par défaut: "Annee" (pour février XXXX). Sinon: février.}

\item{majuscule}{Par défaut 0. 1 pour avoir "Février" plutôt que "février"}

\item{k}{Par défaut 1 pour le mois suivant. Un nombre entier relatif.}
}
\value{
Février XXXX (ou une variante)
}
\description{
Fournit dynamiquement le trimestre suivant sous forme littéraire
}
\examples{
nextMois(3,2023)                  #avril 2023
nextMois(12,2023)                 #janvier 2024
nextMois(12,2023,k=2)             #février 2024

}
\seealso{
quelMois nextTrim prevMois
}
