% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quelTrim.R
\name{nextTrim}
\alias{nextTrim}
\title{Fournit dynamiquement le trimestre suivant}
\usage{
nextTrim(
  trim,
  annee,
  type = c("lettres", "chiffres"),
  majuscule = FALSE,
  exposant = TRUE,
  k = 1
)
}
\arguments{
\item{trim}{Trimestre sous forme numerique : 1, 2, 3 ou 4.}

\item{annee}{Annee sur 4 chiffres.}

\item{type}{"lettres" (par defaut) ou "chiffres".}

\item{majuscule}{Indicateur logique : TRUE pour une majuscule initiale,
FALSE sinon.}

\item{exposant}{Indicateur logique : TRUE pour afficher
les exposants Markdown (ex. "1^er^"),
FALSE pour afficher "1er".}

\item{k}{Decalage en nombre de trimestres.
Par defaut : 1 (trimestre suivant).}
}
\value{
Une chaine de caracteres correspondant au trimestre
suivant (par exemple : "quatrieme trimestre 2023").
}
\description{
Retourne la formulation du trimestre situe k periodes
apres le trimestre indique.
}
\examples{
nextTrim(3, 2023) # "quatrième trimestre 2023"
nextTrim(4, 2023) # "premier trimestre 2024"

}
\seealso{
\code{\link{quelTrim}}, \code{\link{prevTrim}}
}
