% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quelTrim.R
\name{nextTrim}
\alias{nextTrim}
\title{Fournit dynamiquement le trimestre suivant}
\usage{
nextTrim(
  trim,
  annee,
  type = c("lettres", "chiffres"),
  majuscule = FALSE,
  exposant = TRUE,
  k = 1,
  lang = get_serad_language()
)
}
\arguments{
\item{trim}{Trimestre sous forme numérique : 1, 2, 3 ou 4.}

\item{annee}{Année sur 4 chiffres.}

\item{type}{"lettres" (par défaut) ou "chiffres".}

\item{majuscule}{Indicateur logique : TRUE pour une majuscule initiale,
FALSE sinon.}

\item{exposant}{Indicateur logique : TRUE pour afficher
les exposants Markdown (ex. "1^er^"),
FALSE pour afficher "1er".}

\item{k}{Décalage en nombre de trimestres.
Par défaut : 1 (trimestre suivant).}

\item{lang}{Langue de sortie : "fr" ou "en".}
}
\value{
Une chaîne de caractères correspondant au trimestre suivant.
}
\description{
Retourne la formulation du trimestre situé à \code{k} périodes
après le trimestre indiqué.
}
\examples{
nextTrim(3, 2023)
nextTrim(4, 2023)

}
\seealso{
\code{\link{quelTrim}}, \code{\link{prevTrim}}
}
