% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plushautniveau.R
\name{plushautniveau}
\alias{plushautniveau}
\title{Indique si le niveau est le plus haut ou le plus bas de la serie}
\usage{
plushautniveau(
  df,
  temps = "croissant",
  voc_haut = "C'est le plus haut niveau depuis",
  voc_bas = "C'est le plus bas niveau depuis",
  vart,
  vary,
  nbperiode = 3
)
}
\arguments{
\item{df}{Data frame en entree.}

\item{temps}{Ordre temporel : "croissant" (par defaut)
si les valeurs vont du passe vers le present,
"decroissant" sinon.}

\item{voc_haut}{Formulation utilisee pour un plus haut niveau.}

\item{voc_bas}{Formulation utilisee pour un plus bas niveau.}

\item{vart}{Variable de temps du data frame.
Par defaut : premiere colonne.}

\item{vary}{Variable numerique du data frame.
Par defaut : seconde colonne.}

\item{nbperiode}{Nombre minimal de periodes depuis
lesquelles le niveau doit etre extremum pour
retourner une phrase (par defaut : 3).}
}
\value{
Une chaine de caracteres indiquant le plus haut
ou le plus bas niveau depuis une date donnee.
Retourne une chaine vide si rien de notable.
}
\description{
Determine si la derniere valeur d'une serie correspond
au plus haut ou au plus bas niveau observe depuis
un certain nombre de periodes.
}
\examples{
col0 = c("Y1T1", "Y1T2", "Y1trim3", "Y1T4","Y2T1","Y2-T2")
col1 = c(12, 11, 7, 6, 9, 10)
col2 = c(12, 11, 7, 6, 9, 14)
col3 = c(12, 11, 3, 6, 9, 4)
col4 = c(12, 11, 7, 6, 9, 4)
col5 = c(12, 11, 7, 3, 9, 4)
df1 = data.frame(col0, col1, col2, col3, col4, col5)

plushautniveau(df1)
# "C'est le plus haut niveau depuis Y1T2."

plushautniveau(df1, nbperiode = 5)
# ""

plushautniveau(df1, vary = "col2")
# "C'est le plus haut niveau depuis le debut de la serie."

plushautniveau(df1, vary = "col3")
# "C'est le plus bas niveau depuis Y1trim3."

plushautniveau(df1, vary = "col4")
# "C'est le plus bas niveau depuis le debut de la serie."

}
