% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plushautniveau.R
\name{plushautniveau}
\alias{plushautniveau}
\title{Précise si le plus haut/niveau de la série et depuis quelle date}
\usage{
plushautniveau(
  df,
  temps = "croissant",
  voc_haut = "C'est le plus haut niveau depuis",
  voc_bas = "C'est le plus bas niveau depuis",
  vart,
  vary,
  nbperiode = 3
)
}
\arguments{
\item{df}{Le dataframe en entrée}

\item{temps}{"croissant" par défaut si du passé vers le présent. "décroissant" sinon.}

\item{voc_haut}{"C'est le plus haut niveau de la série depuis" par défaut}

\item{voc_bas}{"C'est le plus bas niveau de la série depuis" par défaut}

\item{vart}{La variable de temps du dataframe. La première valeur par défaut.}

\item{vary}{La variable numérique du dataframe. La seconde valeur par défaut.}

\item{nbperiode}{Le nombre de périodes dépassée à partir duquel c'est utile
de sortir une valeur. 3 par défaut.}
}
\value{
La phrase. Rien si rien de notable.
}
\description{
Un dataframe en entrée.\cr
Si pas indiqué, la date est considérée comme étant la première colonne.
Pas besoin de la formater.\cr
Si pas indiqué, on suppose les valeurs les plus anciennes en haut
(temps croissant).\cr
Si pas indiqué, la série d'intérêt, numérique, est considérée comme étant la
seconde colonne.\cr
Sauf si le paramètre diffère, la fonction ne ressort rien si pas la plus
haute/basse valeur depuis au moins 3 périodes.
}
\details{
Fonction ecrite sur demande d'un utilisateur.
}
