% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quelTrim.R
\name{prevMois}
\alias{prevMois}
\title{Fournit dynamiquement le mois precedent}
\usage{
prevMois(mois, annee, type = c("Annee", "autres"), majuscule = FALSE, k = 1)
}
\arguments{
\item{mois}{Mois sous forme numerique : 1 a 12.}

\item{annee}{Annee sur 4 chiffres.}

\item{type}{"Annee" (par defaut, ex. "decembre 2023")
ou "autres" (ex. "decembre").}

\item{majuscule}{Indicateur logique : TRUE pour une majuscule
initiale (ex. "Decembre"), FALSE sinon.}

\item{k}{Decalage en nombre de mois.
Par defaut : 1 (mois precedent).}
}
\value{
Une chaine de caracteres correspondant au mois
precedent (par exemple : "decembre 2022").
}
\description{
Retourne la formulation du mois situe k periodes
avant le mois indique.
}
\examples{
prevMois(1, 2023)         # "décembre 2022"
prevMois(12, 2023, k = 2) # "octobre 2023"

}
\seealso{
\code{\link{prevTrim}}, \code{\link{nextMois}}, \code{\link{quelMois}}
}
