% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quelTrim.R
\name{prevMois}
\alias{prevMois}
\title{Fournit dynamiquement le mois précédent}
\usage{
prevMois(
  mois,
  annee,
  type = c("Annee", "autres"),
  majuscule = FALSE,
  k = 1,
  lang = get_serad_language()
)
}
\arguments{
\item{mois}{Mois sous forme numérique : 1 à 12.}

\item{annee}{Année sur 4 chiffres.}

\item{type}{"Annee" (par défaut) ou "autres".}

\item{majuscule}{Indicateur logique : TRUE pour une majuscule
initiale, FALSE sinon.}

\item{k}{Décalage en nombre de mois.
Par défaut : 1 (mois précédent).}

\item{lang}{Langue de sortie : "fr" ou "en".}
}
\value{
Une chaîne de caractères correspondant au mois précédent.
}
\description{
Retourne la formulation du mois situé à \code{k} périodes
avant le mois indiqué.
}
\examples{
prevMois(1, 2023)
prevMois(12, 2023, k = 2)

}
\seealso{
\code{\link{prevTrim}}, \code{\link{nextMois}}, \code{\link{quelMois}}
}
