% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quelTrim.R
\name{prevMois}
\alias{prevMois}
\title{Fournit dynamiquement le mois précédent sous forme littéraire}
\usage{
prevMois(mois, annee, type = "Annee", majuscule = 0, k = 1)
}
\arguments{
\item{mois}{Le mois sous la forme d'un chiffre : 1..12}

\item{annee}{Un nombre positif en 4 chiffres}

\item{type}{Par défaut: "Annee" (pour décembre XXXX). Sinon: décembre.}

\item{majuscule}{Par défaut 0. 1 pour avoir "Décembre" plutôt que "décembre"}

\item{k}{Par défaut 1 pour le mois précédent. Un nombre entier relatif.}
}
\value{
décembre XXXX (ou une variante)
}
\description{
Fournit dynamiquement le mois précédent sous forme littéraire
}
\examples{
prevMois(1,2023)                  #décembre 2022
prevMois(12,2023,k=2)             #octobre  2023

}
\seealso{
prevTrim nextMois quelMois
}
