% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quelTrim.R
\name{quelMois}
\alias{quelMois}
\title{Fournit dynamiquement le mois en toutes lettres}
\usage{
quelMois(mois, annee, type = c("Annee", "autres"), majuscule = FALSE)
}
\arguments{
\item{mois}{Mois sous forme numerique : 1 a 12.}

\item{annee}{Annee sur 4 chiffres.}

\item{type}{"Annee" (par defaut, ex. "janvier 2023")
ou "autres" (ex. "janvier").}

\item{majuscule}{Indicateur logique : TRUE pour une majuscule
initiale (ex. "Janvier"), FALSE sinon.}
}
\value{
Une chaine de caracteres correspondant au mois,
avec ou sans annee.
}
\description{
Retourne une formulation litteraire du mois,
avec ou sans annee.
}
\examples{
quelMois(3, 2023)
quelMois(3, 2023, majuscule = TRUE)
quelMois(3, 2023, type = "autres")

}
\seealso{
\code{\link{quelTrim}}, \code{\link{nextMois}}, \code{\link{prevMois}}
}
