% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quelTrim.R
\name{quelMois}
\alias{quelMois}
\title{Fournit dynamiquement le mois en toutes lettres}
\usage{
quelMois(
  mois,
  annee,
  type = c("Annee", "autres"),
  majuscule = FALSE,
  lang = get_serad_language()
)
}
\arguments{
\item{mois}{Mois sous forme numérique : 1 à 12.}

\item{annee}{Année sur 4 chiffres.}

\item{type}{"Annee" (par défaut, ex. "janvier 2023")
ou "autres" (ex. "janvier").}

\item{majuscule}{Indicateur logique : TRUE pour une majuscule
initiale (ex. "Janvier"), FALSE sinon.}

\item{lang}{Langue de sortie : "fr" ou "en".}
}
\value{
Une chaîne de caractères correspondant au mois,
avec ou sans année.
}
\description{
Retourne une formulation littéraire du mois,
avec ou sans année.
}
\examples{
quelMois(3, 2023)
quelMois(3, 2023, majuscule = TRUE)
quelMois(3, 2023, type = "autres")

}
\seealso{
\code{\link{quelTrim}}, \code{\link{nextMois}}, \code{\link{prevMois}}
}
