% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quelTrim.R
\name{quelMois}
\alias{quelMois}
\title{Fournit dynamiquement le mois sous forme littéraire}
\usage{
quelMois(mois, annee, type = "Annee", majuscule = 0)
}
\arguments{
\item{mois}{Le mois sous la forme d'un chiffre : 1...12}

\item{annee}{Un nombre positif en 4 chiffres}

\item{type}{Par défaut: "Annee" (pour janvier XXXX). Sinon: janvier}

\item{majuscule}{Par défaut 0. 1 pour avoir "Janvier" plutôt que "janvier"}
}
\value{
janvier XXXX (ou une variante)
}
\description{
Fournit dynamiquement le mois sous forme littéraire
}
\examples{
quelMois(3,2023)                  #mars 2023
quelMois(3,2023,majuscule=1)      #Mars 2023
quelMois(3,2023,type="autres")    #mars

}
\seealso{
quelTrim nextMois prevMois
}
