% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{s}
\alias{s}
\title{Met au pluriel selon la valeur numerique}
\usage{
s(a, sing = "", plur = "s", seuil = 2)
}
\arguments{
\item{a}{Valeur numerique.}

\item{sing}{Forme au singulier. Par defaut : chaine vide.}

\item{plur}{Forme au pluriel. Par defaut : "s".}

\item{seuil}{Seuil a partir duquel la forme plurielle
est utilisee (par defaut : 2).}
}
\value{
Une chaine de caracteres correspondant a la forme
correctement accordee. NA si l'entree est NA.
}
\description{
Retourne la forme singuliere ou plurielle
selon la valeur absolue du nombre.
}
\examples{
serad::s(-7.5)
serad::s(-2)
serad::s(1.4, "chat parle", "chats parlent")
serad::s(-2, "chat parle", "chats parlent")
serad::s(1.97)
serad::s(arrondi_tot(1.97))
serad::s(1.97, seuil = 1.95)

}
\seealso{
\url{https://www.dictionnaire-academie.fr/article/QDL057}
}
