% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{s}
\alias{s}
\title{Met au pluriel si la valeur absolue du nombre est supérieure ou égale a 2.}
\usage{
s(a, sing = "", plur = "s", seuil = 1.95)
}
\arguments{
\item{a}{Le nombre}

\item{sing}{Le mot à indiquer au singulier, rien par défaut.}

\item{plur}{Le mot à indiquer au pluriel, -s par défaut (d'ou le nom de la fonction)}

\item{seuil}{Pour gérer le cas limite où un arrondi à 2 depuis l'inférieur.}
}
\value{
Le mot bien accordé ; par défault '-s ou rien.
}
\description{
Met au pluriel si la valeur absolue du nombre est supérieure ou égale a 2.
}
\examples{
s(-7.5)
s(-2)
s(1.4,"chat parle", "chats parlent")
s(-2,"chat parle", "chats parlent")

}
