% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quelTrim.R
\name{whichMois}
\alias{whichMois}
\title{Recupere le numero du mois a partir d'un texte}
\usage{
whichMois(mois)
}
\arguments{
\item{mois}{Chaine de caracteres contenant un mois
(ex. "En Juil 98").}
}
\value{
Un entier entre 1 et 12 si un mois est detecte,
0 sinon. NA si l'entree est NA.
}
\description{
Extrait le numero du mois (1 a 12) a partir
d'une chaine de caracteres contenant un mois
en format abrege ou litteraire.
}
\examples{
whichMois("En Juil 98")

}
\seealso{
\code{\link{quelMois}}
}
