% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quelTrim.R
\name{whichMois}
\alias{whichMois}
\title{Récupère le numéro du mois à partir d'un texte}
\usage{
whichMois(mois)
}
\arguments{
\item{mois}{Chaîne de caractères contenant un mois
(ex. "En juil. 98").}
}
\value{
Un entier entre 1 et 12 si un mois est détecté,
0 sinon. NA si l'entrée est NA.
}
\description{
Extrait le numéro du mois (1 à 12) à partir
d'une chaîne de caractères contenant un mois
en format abrégé ou littéraire.
}
\examples{
whichMois("En juil. 98")

}
\seealso{
\code{\link{quelMois}}
}
