% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quelTrim.R
\name{whichMois}
\alias{whichMois}
\title{pour récupérer le mois dans un format littéraire}
\usage{
whichMois(mois)
}
\arguments{
\item{mois}{Le mois en lettres}
}
\value{
1..12
}
\description{
pour récupérer le mois dans un format littéraire
}
\examples{
whichMois("En Juil 98")                  #7

}
\seealso{
quelMois
}
